feat: ボード作成モーダルのflagを使用したいコンポーネントが多い為、ボード作成モーダルのflagをinitialStateで管理するよう修正()
