+++ src/components/CreateNewBoard.tsx
@@ -5,6 +5,7 @@ import Button from '@material-ui/core/Button';
 import Paper from '@material-ui/core/Paper';
 import CreateNewBoardModal from 'components/CreateNewBoardModal';
 import AppContext from 'contexts/AppContext';
+import { SEND_OPEN_FLAG } from 'actions/index';
 
 const useStyles = makeStyles((theme: Theme) =>
   createStyles({
@@ -22,22 +23,6 @@ const useStyles = makeStyles((theme: Theme) =>
   })
 );
 
-const boardStyle = {
-  backgroundColor: '#0387c7',
-  width: 345,
-  margin: 20,
-  padding: '25px 0',
-  borderRadius: 5.5,
-};
-
-const boardStyleAddMarginTop = {
-  backgroundColor: '#0387c7',
-  width: 345,
-  margin: '200px 20px 20px 20px ',
-  padding: '25px 0',
-  borderRadius: 5.5,
-};
-
 const buttonStyle = {
   backgroundColor: '#9fe7a4',
   width: 345,
@@ -47,77 +32,34 @@ const buttonStyle = {
 };
 
 const CreateNewBoard: React.FC = () => {
-  const [anchorEl, setAnchorEl] = React.useState<HTMLButtonElement | null>(
-    null
-  );
-  const [open, setOpen] = React.useState(false);
-  const [placement, setPlacement] = React.useState<PopperPlacementType>();
+  const [anchorEl, setAnchorEl] = useState<HTMLButtonElement | null>(null);
+  const [placement, setPlacement] = useState<PopperPlacementType>();
   const classes = useStyles();
 
-  const { state } = useContext(AppContext);
-  const [modalFlag, setModalFlag] = useState(false);
-
+  const { state, dispatch } = useContext(AppContext);
   const handleClick = (newPlacement: PopperPlacementType) => (
     event: React.MouseEvent<HTMLButtonElement>
   ) => {
     setAnchorEl(event.currentTarget);
-    setOpen(prev => placement !== newPlacement || !prev);
     setPlacement(newPlacement);
-    setModalFlag(!modalFlag);
+    dispatch({ type: SEND_OPEN_FLAG, modalFlag: state.openFlag });
   };
-  /*
-    モーダル開いてる時
-    2行目の配列だけにmarginTopを付与する為、配列を切り分ける処理
-  */
-  let firstLineArray = [];
-  let secondLineArray = [];
-  let afterThirdLine = [];
-  if (window.innerWidth > 1154) {
-    firstLineArray = state.createdBordArray.slice(0, 2);
-    secondLineArray = state.createdBordArray.slice(2, 5);
-    afterThirdLine = state.createdBordArray.slice(5);
-  }
-  if (1155 > window.innerWidth && window.innerWidth > 770) {
-    firstLineArray = state.createdBordArray.slice(0, 1);
-    secondLineArray = state.createdBordArray.slice(1, 3);
-    afterThirdLine = state.createdBordArray.slice(3);
-  }
-  if (window.innerWidth < 770) {
-    secondLineArray = state.createdBordArray.slice(0, 1);
-    afterThirdLine = state.createdBordArray.slice(1);
-  }
-  const secondLineStyle =
-    modalFlag === true ? boardStyleAddMarginTop : boardStyle;
 
   return (
     <>
       <Button onClick={handleClick('bottom-start')} style={buttonStyle}>
         <h4>Creating a board</h4>
       </Button>
-      <Popper open={open} anchorEl={anchorEl} placement={placement} transition>
+      <Popper
+        open={state.openFlag}
+        anchorEl={anchorEl}
+        placement={placement}
+        transition
+      >
         <Paper className={classes.typography}>
           <CreateNewBoardModal />
         </Paper>
       </Popper>
-      {firstLineArray.map((val: any, index: number) => (
-        <div key={index} style={boardStyle}>
-          <h4 style={{ textAlign: 'center' }}>{val.text}</h4>
-        </div>
-      ))}
-      {secondLineArray.length > 0
-        ? secondLineArray.map((val: any, index: number) => (
-            <div key={index} style={secondLineStyle}>
-              <h4 style={{ textAlign: 'center' }}>{val.text}</h4>
-            </div>
-          ))
-        : ''}
-      {afterThirdLine.length > 0
-        ? afterThirdLine.map((val: any, index: number) => (
-            <div key={index} style={boardStyle}>
-              <h4 style={{ textAlign: 'center' }}>{val.text}</h4>
-            </div>
-          ))
-        : ''}
     </>
   );
 };
