+++ src/reducers/createdBordArray.ts
@@ -3,6 +3,7 @@ import { State } from 'components/App';
 
 const initialState: State = {
   createdBordArray: [],
+  openFlag: false,
 };
 
 const createdBordArray = (
