+++ src/reducers/index.ts
@@ -1,5 +1,6 @@
 import { combineReducers } from 'redux';
 
 import createdBordArray from './createdBordArray';
+import openFlag from './openFlag';
 
-export default combineReducers({ createdBordArray });
+export default combineReducers({ createdBordArray, openFlag });
